// src/app/shared/enums/config-key.enum.ts
export enum ConfigKeyEnum {
  apiBaseUrl = 'apiBaseUrl',
  masterApiBaseUrl = 'masterApiBaseUrl',
  azureBlobUrl = 'azureBlobUrl',
  authClientId = 'authClientId',
  scopes = 'scopes',
  baseAuthority = 'baseAuthority',
  authorityDomain = 'authorityDomain',
  mapApiKey = 'mapApiKey',
  isProduction = 'isProduction',
}